Used longs for shot numbers

+++ mdsplus/javascope/AsdexDataProvider.java
@@ -1,4 +1,4 @@
-/* $Id: AsdexDataProvider.java,v 1.2 2002/04/26 13:14:29 jgk Exp $ */
+/* $Id: AsdexDataProvider.java,v 1.3 2002/07/26 10:55:55 manduchi Exp $ */
 import java.util.*;
 import java.io.IOException;
 import javax.swing.JFrame;
@@ -22,7 +22,7 @@ class AsdexDataProvider extends MdsDataProvider
         mds.setUser("mdsplus");
     }
     
-    public synchronized void Update(String exp, int s)
+    public synchronized void Update(String exp, long s)
 	{
 	    error = null;
 		shot = s;
